requirements.txt
allure-pytest==2.9.43
allure-python-commons==2.9.43
apipkg==1.5
atomicwrites==1.4.0
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
colorama==0.4.4
configparser==5.0.2
crayons==0.4.0
execnet==1.8.1
idna==2.10
iniconfig==1.1.1
packaging==20.9
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
pytest-forked==1.3.0
pytest-html==3.1.1
pytest-metadata==1.11.0
pytest-rerunfailures==10.0
pytest-xdist==2.2.1
requests==2.25.1
selenium==3.141.0
six==1.16.0
toml==0.10.2
urllib3==1.26.5
webdriver-manager==3.4.2
